visualization/requirements.txt
pandas
matplotlib
sqlalchemy
psycopg2-binary
python-dotenv